Split tests into files (one per test subject)
